Updated Apigee.MobileAnalytics to use URI param
Apigee.MobileAnalytics was looking for 'apiURL' in the options object, while Usergrid.Client was looking for 'URI'. So if a user wanted to override the default prod server, they had to specify two different params to get to the same base url for each object.

I've changed Apigee.MobileAnalytics to look for URI, and appended the trailing slash to all instances where apiURL was being used to form a URL (so that it acts the same as Usergrid.Client). Also removed the trailing slash from the default base url of Apigee.MobileAnalytics.